requirements.txt
pyserial
simpleaudio
paho-mqtt
requests